Preventing nasty accidents

+++ .lando/scripts/delete_database.sh
@@ -16,8 +16,12 @@ case $i in
 esac
 done
 
-lando_green "Deleting database: ${database}";
+# Ensure we are not deleting anything critical!
+if { [ $database = 'mysql' ] || [ $database = 'information_schema' ] || [ $database = 'performance_schema' ]; }; then
+  lando_red "I'm sorry Dave, I'm afraid I can't do that"
+else
+  lando_green "Deleting database: ${database}";
 
-# Drop the database if it exists.
-mysql -uroot -e "DROP DATABASE IF EXISTS ${database}";
-
+  # Drop the database if it exists.
+  mysql -uroot -e "DROP DATABASE IF EXISTS ${database}";
+fi
